+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/enums/core/EscapeType.kt
@@ -2,9 +2,10 @@ package com.windea.breezeframework.core.enums.core
 
 /**转义类型。*/
 enum class EscapeType(
-	val escapeStrings: Array<String>, //ignore "\"
-	val escapedStrings: Array<String>,
-	val escapeBackslash: Boolean = true
+	/**转义的字符串组。*/
+	val escapeStrings: Array<String>,
+	/**转义后的字符串组。*/
+	val escapedStrings: Array<String>
 ) {
 	/**Kotlin转义。*/
 	Kotlin(
@@ -16,15 +17,14 @@ enum class EscapeType(
 		arrayOf("\t", "\b", "\n", "\r", "\'", "\""),
 		arrayOf("\\t", "\\b", "\\n", "\\r", "\\'", "\\\"")
 	),
-	/**正则表达式转义。*/
+	/**正则表达式转义。注意：不是将原始字符串用`\Q` `\E`包围，而是转义其中必要的字符。*/
 	Regex(
-		arrayOf(".", "^", "$", "[", "{", "(", "|", "?", "+"), //not for every scope
-		arrayOf("\\.", "\\^", "\\$", "\\[", "\\{", "\\(", "\\|", "\\?", "\\+")
+		arrayOf(".", "^", "$", "[", "{", "(", "|", "-", "*", "?", "+"), //not for every scope
+		arrayOf("\\.", "\\^", "\\$", "\\[", "\\{", "\\(", "\\|", "-", "\\*", "\\?", "\\+")
 	),
 	/**Xml转义。*/
 	Xml(
 		arrayOf("<", ">", "&", "'", "\""), //only for basic escape
-		arrayOf("&lt;", "&gt;", "&amp;", "&apos;", "quot;"),
-		false
+		arrayOf("&lt;", "&gt;", "&amp;", "&apos;", "quot;")
 	)
 }