+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/enums/core/FormatCase.kt
@@ -2,7 +2,12 @@ package com.windea.breezeframework.core.enums.core
 
 /**显示格式。*/
 interface FormatCase {
-	val regex: Regex
+	/**分割方法。*/
 	val splitFunction: (String) -> List<String>
+	/**加入方法。*/
 	val joinFunction: (List<String>) -> String
+	/**用于验证合法性的正则表达式。*/
+	val regex: Regex?
+	/**用于验证合法性的预测。*/
+	val predicate: (String) -> Boolean
 }