+++ breeze-dsl/breeze-dsl-puml/src/main/kotlin/com/windea/breezeframework/dsl/puml/PumlStateDiagramDsl.kt
@@ -197,7 +197,7 @@ class PumlStateDiagramConcurrentSection : PumlStateDiagramDslEntry {
 }
 
 /**
- * PlantUml状态图转换。
+ * PlantUml状态图转化。
  *
  * 语法示例：
  * * `A -> B`